fix: improve some names
